Support `new-only` flag
Instead of getting the new posts based on the number of weeks requested,
we take the last time `gobbler` was run, and get the posts since then.
This doesn't work completely correctly yet, as `get_last_ran_at` (and in
fact all the date time related stuff) ignores the time component. This
has to be fixed for this feature to work correctly.