Offline support for `--run-days`
When offline, no posts will be found. If `--run-days` is passed,
`gobbler` will not run for the rest of the specified time, even though
an internet connection may be established in the mean time, which could
yield new results. By not saving whether we ran today if we didn't find
any posts, we can ensure that we will try again when we're reconnected.